+++ apps/dashboard/src/routes/organizations/[id]/+page.svelte
@@ -85,6 +85,24 @@
 			goto('/');
 		}
 	};
+
+	async function subscribe(priceId: string) {
+		const res = await fetch('/api/stripe/subscribe', {
+			method: 'POST',
+			body: JSON.stringify({ priceId: priceId })
+		});
+		const { url } = await res.json();
+		goto(url);
+	}
+
+	async function manage(subscriptionId: string) {
+		const res = await fetch('/api/stripe/manage', {
+			method: 'POST',
+			body: JSON.stringify({ subscriptionId: subscriptionId })
+		});
+		const { url } = await res.json();
+		goto(url);
+	}
 </script>
 
 <div class="sm:container sm:mx-auto justify-center p-8">
@@ -156,4 +174,29 @@
 			</div>
 		</div>
 	</div>
+
+	<!-- WIP -->
+	{#if !data.activeSubscription}
+		{#each data.subscriptionPlans as subscription}
+			<button
+				type="button"
+				class="btn variant-ghost-primary"
+				on:click={() => {
+					subscribe(subscription.id);
+				}}
+			>
+				<span>{subscription.billing_products?.name}</span>
+			</button>
+		{/each}
+	{:else}
+		<button
+			type="button"
+			class="btn variant-ghost-primary"
+			on:click={() => {
+				if (data.activeSubscription) manage(data.activeSubscription.id);
+			}}
+		>
+			<span>Modify active subscription</span>
+		</button>
+	{/if}
 </div>
